Split real Pi lifecycle budget tests
