+++ tests2/core/tool-result-error-bridge-real-pi.test.ts
@@ -1,197 +1,28 @@
 import assert from "node:assert/strict";
 import fs from "node:fs";
-import os from "node:os";
 import path from "node:path";
 
-import { Agent } from "@earendil-works/pi-agent-core";
-import { createAssistantMessageEventStream } from "@earendil-works/pi-ai";
-import {
-	AgentSession,
-	ExtensionRunner,
-	SessionManager,
-	SettingsManager,
-	type LoadExtensionsResult,
-} from "@earendil-works/pi-coding-agent";
-// Pi 0.81's runtime root accidentally omits this export although its declaration
-// file includes it. Import the installed loader directly so this remains a real
-// loader regression rather than another synthetic ExtensionAPI fixture.
-import { loadExtensions } from "../../node_modules/@earendil-works/pi-coding-agent/dist/core/extensions/loader.js";
 import { afterEach, describe, it } from "vitest";
 
 import {
+	bytes,
+	cleanupRealPiRoots,
+	createLifecycleSession,
 	generateToolResultErrorBridgeExtension,
+	loadExtensions,
+	makeRealPiRoot,
+	persistOutcome,
 	READ_SESSION_FINAL_RESULT_MAX_BYTES,
-} from "../../src/server/agent/tool-result-error-bridge-extension.js";
+	resultValueFromMessage,
+	runLifecycle,
+	toolOutcome,
+} from "./tool-result-error-bridge-real-pi-fixture.js";
 
-const roots: string[] = [];
-
-const usage = {
-	input: 0,
-	output: 0,
-	cacheRead: 0,
-	cacheWrite: 0,
-	totalTokens: 0,
-	cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0, total: 0 },
-};
-
-function model() {
-	return {
-		id: "test-model",
-		name: "Test model",
-		api: "test",
-		provider: "test",
-		baseUrl: "https://example.invalid",
-		reasoning: false,
-		input: ["text"],
-		cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 },
-		contextWindow: 100_000,
-		maxTokens: 1_000,
-	} as any;
-}
-
-interface StreamToolCall {
-	name: string;
-	arguments: Record<string, unknown>;
-	id: string;
-}
-
-function makeSequenceStream(calls: StreamToolCall[]) {
-	let index = 0;
-	return () => {
-		const stream = createAssistantMessageEventStream();
-		const next = calls[index++];
-		const message = next
-			? {
-				role: "assistant" as const,
-				content: [{ type: "toolCall" as const, id: next.id, name: next.name, arguments: next.arguments }],
-				api: "test",
-				provider: "test",
-				model: "test-model",
-				usage,
-				stopReason: "toolUse" as const,
-				timestamp: Date.now(),
-			}
-			: {
-				role: "assistant" as const,
-				content: [{ type: "text" as const, text: "done" }],
-				api: "test",
-				provider: "test",
-				model: "test-model",
-				usage,
-				stopReason: "stop" as const,
-				timestamp: Date.now(),
-			};
-		stream.push({ type: "start", partial: message });
-		stream.push({ type: "done", reason: message.stopReason, message });
-		return stream;
-	};
-}
-
-function makeStream(toolName: string, args: Record<string, unknown>, toolCallId?: string) {
-	return makeSequenceStream([{ name: toolName, arguments: args, id: toolCallId ?? "call-1" }]);
-}
-
-function resourceLoader(loaded: LoadExtensionsResult): any {
-	return {
-		getExtensions: () => loaded,
-		getSkills: () => ({ skills: [], diagnostics: [] }),
-		getPrompts: () => ({ prompts: [], diagnostics: [] }),
-		getThemes: () => ({ themes: [], diagnostics: [] }),
-		getAgentsFiles: () => ({ agentsFiles: [] }),
-		getSystemPrompt: () => undefined,
-		getAppendSystemPrompt: () => [],
-		extendResources: () => undefined,
-		reload: async () => undefined,
-	};
-}
-
-let lifecycleCounter = 0;
-
-function createLifecycleSession(
-	loaded: LoadExtensionsResult,
-	root: string,
-	args: Record<string, unknown>,
-	toolName = "read_session",
-	toolCallId?: string,
-	streamFn = makeStream(toolName, args, toolCallId),
-): AgentSession {
-	const sessionManager = SessionManager.create(root, path.join(root, `lifecycle-${++lifecycleCounter}`));
-	const settingsManager = SettingsManager.inMemory({
-		compaction: { enabled: false },
-		retry: { enabled: false, maxRetries: 0 },
-	});
-	const agent = new Agent({
-		initialState: {
-			systemPrompt: "test",
-			model: model(),
-			tools: [],
-		},
-		streamFn,
-	});
-	return new AgentSession({
-		agent,
-		sessionManager,
-		settingsManager,
-		cwd: root,
-		resourceLoader: resourceLoader(loaded),
-		modelRuntime: {} as never,
-		baseToolsOverride: {},
-		initialActiveToolNames: [],
-	});
-}
-
-async function runLifecycle(session: AgentSession, label: string): Promise<any[]> {
-	const events: any[] = [];
-	session.subscribe((event) => { events.push(event); });
-	await (session as any)._runAgentPrompt({
-		role: "user",
-		content: [{ type: "text", text: label }],
-		timestamp: Date.now(),
-	});
-	return events;
-}
-
-function bytes(value: unknown): number {
-	return Buffer.byteLength(JSON.stringify(value), "utf8");
-}
-
-function toolOutcome(events: any[]) {
-	const emitted = events.find((event) => event.type === "tool_execution_end");
-	const persisted = events.find((event) => event.type === "message_end" && event.message.role === "toolResult")?.message;
-	assert.ok(emitted);
-	assert.ok(persisted);
-	return { emitted, persisted };
-}
-
-function persistOutcome(session: AgentSession) {
-	const sessionFile = session.sessionManager.getSessionFile();
-	assert.ok(sessionFile);
-	const roundTrip = (SessionManager.open(sessionFile).getEntries()
-		.find((entry) => entry.type === "message" && entry.message.role === "toolResult") as any)?.message;
-	assert.ok(roundTrip);
-	const line = fs.readFileSync(sessionFile, "utf8")
-		.split(/\r?\n/)
-		.find((candidate) => candidate.includes('"role":"toolResult"'));
-	assert.ok(line);
-	return { roundTrip, line, sessionFile };
-}
-
-function resultValueFromMessage(message: any): Record<string, unknown> {
-	const value: Record<string, unknown> = {};
-	for (const key of ["content", "details", "isError"] as const) {
-		if (Object.prototype.hasOwnProperty.call(message, key)) value[key] = message[key];
-	}
-	return value;
-}
-
-afterEach(() => {
-	for (const root of roots.splice(0)) fs.rmSync(root, { recursive: true, force: true });
-});
+afterEach(cleanupRealPiRoots);
 
 describe("tool result boundary through Pi's real extension runner", () => {
 	it("reprojects in-place post-listener mutations before emission and persistence", async () => {
-		const root = fs.mkdtempSync(path.join(os.tmpdir(), "bobbit-real-pi-result-boundary-"));
-		roots.push(root);
+		const root = makeRealPiRoot("bobbit-real-pi-result-boundary-");
 		const boundaryPath = path.join(root, "boundary.ts");
 		const overridePath = path.join(root, "override.ts");
 		const inPlaceMutatorPath = path.join(root, "in-place-mutator.ts");
@@ -889,239 +720,5 @@ export default function (pi) {
 			}
 			assert.equal(stored.roundTrip.toolCallId, normalizedId);
 		}
-	});
-
-	it("does not let cached IDs capture an explicitly named non-read tool", async () => {
-		const root = fs.mkdtempSync(path.join(os.tmpdir(), "bobbit-real-pi-result-collision-"));
-		roots.push(root);
-		const boundaryPath = path.join(root, "boundary.ts");
-		const toolsPath = path.join(root, "tools.ts");
-		fs.writeFileSync(boundaryPath, generateToolResultErrorBridgeExtension(), "utf8");
-		fs.writeFileSync(toolsPath, `
-import { Type } from "typebox";
-
-export default function (pi) {
-  pi.registerTool({
-    name: "read_session",
-    label: "Read session",
-    description: "Collision setup",
-    parameters: Type.Object({ session_id: Type.String() }),
-    async execute() {
-      const envelope = {
-        total: 1,
-        returned: 1,
-        offsetStart: 0,
-        offsetEnd: 0,
-        messages: [{ index: 0, role: "assistant", text: "bounded read" }],
-      };
-      return { content: [{ type: "text", text: JSON.stringify(envelope) }] };
-    },
-  });
-  pi.registerTool({
-    name: "collision_probe",
-    label: "Collision probe",
-    description: "Must remain ordinary tool output",
-    parameters: Type.Object({}),
-    async execute() {
-      return {
-        content: [{ type: "text", text: "NON_READ_COLLISION_PASSTHROUGH" }],
-        details: { providerMetadata: "NON_READ_COLLISION_DETAILS" },
-        usage: { providerMetadata: "NON_READ_COLLISION_USAGE" },
-      };
-    },
-  });
-}
-`, "utf8");
-
-		const loaded = await loadExtensions([boundaryPath, toolsPath], root);
-		assert.deepEqual(loaded.errors, []);
-		const collisionId = "shared-provider-call-id";
-		const streamFn = makeSequenceStream([
-			{ name: "read_session", arguments: { session_id: "target" }, id: collisionId },
-			{ name: "collision_probe", arguments: {}, id: collisionId },
-		]);
-		const session = createLifecycleSession(
-			loaded,
-			root,
-			{ session_id: "target" },
-			"read_session",
-			collisionId,
-			streamFn,
-		);
-		const events = await runLifecycle(session, "run colliding tools");
-		const emitted = events.find((event) =>
-			event.type === "tool_execution_end" && event.toolName === "collision_probe");
-		assert.ok(emitted);
-		assert.equal(emitted.result.content[0].text, "NON_READ_COLLISION_PASSTHROUGH");
-		assert.deepEqual(emitted.result.details, { providerMetadata: "NON_READ_COLLISION_DETAILS" });
-		assert.deepEqual(emitted.result.usage, { providerMetadata: "NON_READ_COLLISION_USAGE" });
-
-		const stateResult = session.state.messages.find((message) =>
-			message.role === "toolResult" && (message as any).toolName === "collision_probe") as any;
-		assert.ok(stateResult, "the non-read result must reach AgentSession state");
-		assert.equal(stateResult.toolCallId, collisionId);
-		assert.equal(stateResult.content[0].text, "NON_READ_COLLISION_PASSTHROUGH");
-		assert.deepEqual(stateResult.details, { providerMetadata: "NON_READ_COLLISION_DETAILS" });
-
-		const sessionFile = session.sessionManager.getSessionFile();
-		assert.ok(sessionFile);
-		const persistedLine = fs.readFileSync(sessionFile, "utf8")
-			.split(/\r?\n/)
-			.find((line) => line.includes('"toolName":"collision_probe"'));
-		assert.ok(persistedLine, "the non-read result must be persisted");
-		assert.equal(persistedLine.includes("NON_READ_COLLISION_PASSTHROUGH"), true);
-		assert.equal(persistedLine.includes("NON_READ_COLLISION_DETAILS"), true);
-	});
-
-	it("retains only fixed-bounded digests for many oversized provider call IDs", async () => {
-		const root = fs.mkdtempSync(path.join(os.tmpdir(), "bobbit-real-pi-result-call-map-"));
-		roots.push(root);
-		const boundaryPath = path.join(root, "boundary.ts");
-		fs.writeFileSync(boundaryPath, generateToolResultErrorBridgeExtension(), "utf8");
-		const loaded = await loadExtensions([boundaryPath], root);
-		assert.deepEqual(loaded.errors, []);
-		const runner = new ExtensionRunner(loaded.extensions, loaded.runtime, root, {} as never, {} as never);
-		const diagnosticsSymbol = Symbol.for("bobbit.tool-result.read-session-call-map-diagnostics.v1");
-		let lastNormalizedId = "";
-
-		for (let index = 0; index < 16; index++) {
-			const prefix = `${index}:`;
-			const providerId = prefix + "x".repeat(2 * 1024 * 1024 - prefix.length);
-			const normalized = await runner.emitMessageEnd({
-				type: "message_end",
-				message: {
-					role: "assistant",
-					content: [{
-						type: "toolCall",
-						id: providerId,
-						name: "read_session",
-						arguments: { session_id: `target-${index}` },
-					}],
-				},
-			} as never) as any;
-			lastNormalizedId = normalized.content[0].id;
-			assert.match(lastNormalizedId, /^brs1:[0-9a-f]{40}$/);
-			assert.notEqual(lastNormalizedId, providerId);
-		}
-
-		const envelope = {
-			total: 1,
-			returned: 1,
-			offsetStart: 0,
-			offsetEnd: 0,
-			messages: [{ index: 0, role: "assistant", text: "correlated" }],
-		};
-		const correlated = await runner.emitToolResult({
-			type: "tool_result",
-			toolCallId: lastNormalizedId,
-			content: [{ type: "text", text: JSON.stringify(envelope) }],
-			isError: false,
-		} as never) as any;
-		assert.equal(JSON.parse(correlated.content[0].text).messages[0].text, "correlated",
-			"the normalized ID must still resolve the correct read_session parameters");
-
-		for (let index = 0; index < 300; index++) {
-			await runner.emitMessageEnd({
-				type: "message_end",
-				message: {
-					role: "assistant",
-					content: [{
-						type: "toolCall",
-						id: `small-${index}`,
-						name: "read_session",
-						arguments: { session_id: `target-${index}` },
-					}],
-				},
-			} as never);
-		}
-
-		const inspect = (runner as any)[diagnosticsSymbol];
-		assert.equal(typeof inspect, "function");
-		const diagnostics = inspect.call(runner);
-		assert.deepEqual(
-			{
-				entries: diagnostics.entries,
-				maxEntries: diagnostics.maxEntries,
-				correlationKeyUnits: diagnostics.correlationKeyUnits,
-				maxKeyUnits: diagnostics.maxKeyUnits,
-			},
-			{ entries: 256, maxEntries: 256, correlationKeyUnits: 45, maxKeyUnits: 45 },
-		);
-		assert.ok(diagnostics.maxValueStringUnits <= 128);
-		assert.ok(diagnostics.totalRetainedStringUnits <= 256 * (45 + 128 + 64 + 64),
-			"all retained map keys and string values must have a fixed aggregate bound");
-	});
-
-	it("always snapshots an unchanged result after the complete listener chain", async () => {
-		const root = fs.mkdtempSync(path.join(os.tmpdir(), "bobbit-real-pi-result-boundary-dedup-"));
-		roots.push(root);
-		const boundaryPath = path.join(root, "boundary.ts");
-		const toolPath = path.join(root, "tool.ts");
-		fs.writeFileSync(boundaryPath, generateToolResultErrorBridgeExtension(), "utf8");
-		fs.writeFileSync(toolPath, `
-import { Type } from "typebox";
-
-export default function (pi) {
-  pi.registerTool({
-    name: "read_session",
-    label: "Read session",
-    description: "Unchanged boundary fixture",
-    parameters: Type.Object({ session_id: Type.String() }),
-    async execute(_toolCallId, params) {
-      const envelope = {
-        total: 1,
-        returned: 1,
-        offsetStart: 0,
-        offsetEnd: 0,
-        messages: [{ index: 0, role: "assistant", text: "bounded once" }],
-      };
-      return {
-        content: [{ type: "text", text: JSON.stringify(envelope) }],
-        details: { session_id: params.session_id, envelope },
-      };
-    },
-  });
-}
-`, "utf8");
-		const loaded = await loadExtensions([boundaryPath, boundaryPath, toolPath], root);
-		assert.deepEqual(loaded.errors, []);
-		const runner = new ExtensionRunner(loaded.extensions, loaded.runtime, root, {} as never, {} as never);
-		const registered = runner.getAllRegisteredTools().find((tool) => tool.definition.name === "read_session");
-		assert.ok(registered);
-		let verboseReads = 0;
-		const input = { session_id: "target" } as Record<string, unknown>;
-		Object.defineProperty(input, "verbose", {
-			get() {
-				verboseReads++;
-				return false;
-			},
-		});
-		const executed = await registered.definition.execute(
-			"call-1",
-			input as never,
-			undefined,
-			undefined as never,
-			{} as never,
-		);
-		assert.equal(verboseReads, 1);
-		assert.equal(Object.getOwnPropertySymbols(executed.details as object).length, 0,
-			"no digest or marker supplied by a mutable result may be trusted");
-
-		const replacement = await runner.emitToolResult({
-			type: "tool_result",
-			toolCallId: "call-1",
-			toolName: "read_session",
-			input,
-			content: executed.content,
-			details: executed.details,
-			isError: false,
-		});
-
-		assert.ok(replacement, "the final seam must always return its own canonical snapshot");
-		assert.equal(verboseReads, 2, "the final seam independently snapshots invocation policy inputs");
-		assert.notEqual(replacement, executed);
-		assert.equal(Object.isFrozen(replacement), true);
-		assert.equal(Object.isFrozen(replacement.content), true);
-		assert.equal(Object.getOwnPropertySymbols(replacement.details as object).length, 0);
 	});
 });
